big oops

+++ src/app/api/post/[postID]/route.ts
@@ -1,7 +1,9 @@
 // API Route
 // /api/post/{postID}
 
+import { Decrypt } from "../../authv2";
 import { DeletePost, GetPost } from "../../db";
+import { cookies } from "next/headers";
 
 // GET /api/post/{postID}
 export async function GET(
@@ -12,7 +14,6 @@ export async function GET(
 
   // get post from parameters
   const postID = (await params).postID;
-  console.log(postID);
 
   const post = await GetPost(postID);
   return new Response(JSON.stringify(post), {
@@ -25,13 +26,40 @@ export async function DELETE(
   request: Request,
   { params }: { params: Promise<{ postID: string }> }
 ) {
-  // TODO: add authentication
+  const session = (await cookies()).get("tkaccess")?.value;
+  // console.log((await cookies()).getAll());
 
-  const postID = (await params).postID;
-  const x = await DeletePost(postID);
-  if (x == true) {
-    return new Response("Success", { status: 200 });
+  // does the session exist?
+  if (!session) {
+    // if no session is present, return 401
+    return new Response(
+      JSON.stringify({ status: "failed", error: "unauthorized1" }),
+      {
+        status: 401,
+      }
+    );
+  }
+  // decrypt session
+  const decrypted_session = await Decrypt(session);
+
+  // is the user of this session an admin?
+  if (decrypted_session.user.role === "admin") {
+    // ok, seems the user was admin, carry on...
+    const postID = (await params).postID;
+    console.log(`deleting post of ID ${postID}...`);
+    const x = await DeletePost(postID);
+    if (x == true) {
+      return new Response("Success", { status: 200 });
+    } else {
+      return new Response("Failed", { status: 500 });
+    }
   } else {
-    return new Response("Failed", { status: 500 });
+    // if user not admin, return 401
+    return new Response(
+      JSON.stringify({ status: "failed", error: "unauthorized2" }),
+      {
+        status: 401,
+      }
+    );
   }
 }
